+++ js/Demo/demoLayers.js
@@ -6,6 +6,7 @@ addLayer("c", {
         name: "Candies", // This is optional, only used in a few places, If absent it just uses the layer id.
         symbol: "C", // This appears on the layer's node. Default is the id with the first letter capitalized
         position: 0, // Horizontal position within a row. By default it uses the layer id and sorts in alphabetical order
+        marked: "discord.png",
         startData() { return {
             unlocked: true,
 			points: new Decimal(0),
@@ -95,7 +96,10 @@ addLayer("c", {
                 rewardDisplay() { return format(this.rewardEffect())+"x" },
                 countsAs: [12, 21], // Use this for if a challenge includes the effects of other challenges. Being in this challenge "counts as" being in these.
                 rewardDescription: "Says hi",
-                onComplete() {console.log("hiii")} // Called when you complete the challenge
+                onComplete() {console.log("hiii")}, // Called when you successfully complete the challenge
+                onEnter() {console.log("So challenging")},
+                onExit() {console.log("Sweet freedom!")},
+
             },
         }, 
         upgrades: {
@@ -315,7 +319,7 @@ addLayer("c", {
                     ["buyables", ""], "blank",
                     ["row", [
                         ["toggle", ["c", "beep"]], ["blank", ["30px", "10px"]], // Width, height
-                        ["display-text", function() {return "Beep"}], "blank", ["v-line", "200px"],
+                        ["layer-proxy", ["f", ["prestige-button"]]], "blank", ["v-line", "200px"],
                         ["column", [
                             ["prestige-button", "", {'width': '150px', 'height': '80px'}],
                             ["prestige-button", "", {'width': '100px', 'height': '150px'}],
@@ -395,7 +399,8 @@ addLayer("f", {
     exponent: 0.5,
     base: 3,
     roundUpCost: true,
-    canBuyMax() {return hasAchievement('a', 13)},
+    canBuyMax() {return false},
+    //directMult() {return new Decimal(player.c.otherThingy)},
 
     row: 1,
     layerShown() {return true}, 
@@ -422,6 +427,7 @@ addLayer("f", {
     canReset() {
         return tmp[this.layer].baseAmount.gte(tmp[this.layer].nextAt)
     },
+
     // This is also non minimal, a Clickable!
     clickables: {
 
@@ -520,5 +526,32 @@ addLayer("a", {
                 onComplete() {console.log("Bork bork bork!")}
             },
         },
-    }, 
+        grid: {
+            maxRows: 3,
+            rows: 2,
+            cols: 2,
+            getStartData(id) {
+                return id
+            },
+            getUnlocked(id) { // Default
+                return true
+            },
+            getCanClick(data, id) {
+                return true
+            },
+            getStyle(data, id) {
+                return {'background-color': '#'+ (data*1234%999999)}
+            },
+            onClick(data, id) { // Don't forget onHold
+                player[this.layer].grid[id]++
+            },
+            getTitle(data, id) {
+                return "Gridable #" + id
+            },
+            getDisplay(data, id) {
+                return data 
+            },
+        } ,
+        midsection: ["grid", "blank"]
+    }
 )