+++ js/game.js
@@ -5,7 +5,7 @@ var scrolled = false;
 
 // Don't change this
 const TMT_VERSION = {
-	tmtNum: "2.5.3",
+	tmtNum: "2.5.5.1",
 	tmtName: "Dreams Really Do Come True"
 }
 
@@ -20,12 +20,13 @@ function getResetGain(layer, canMax=false, useType = null) {
 	} 
 	if(tmp[layer].type == "none")
 		return new Decimal (0)
-	if (tmp[layer].gainExp.eq(0)) return new Decimal(0)
+	if (tmp[layer].gainExp.eq(0)) return decimalZero
 	if (type=="static") {
 		let base = new Decimal(1e8)
 		let exp = new Decimal(1.9)
-		if ((!tmp[layer].canBuyMax) || tmp[layer].baseAmount.lt(tmp[layer].requires)) return new Decimal(1)
+		if ((!tmp[layer].canBuyMax) || tmp[layer].baseAmount.lt(tmp[layer].requires)) return decimalOne
 		let gain = tmp[layer].baseAmount.div(tmp[layer].requires).div(tmp[layer].gainMult).max(1).log(tmp[layer].base).times(tmp[layer].gainExp).pow(Decimal.pow(tmp[layer].exponent, -1))
+		gain = gain.times(tmp[layer].directMult)
 		if (layer == "r"  || layer == "u" || layer == "e") gain = softcapStaticGain(gain, tmp[layer].row)
 		if (layer=="e") gain = gain.mul(tmp.e.infDiv)
 		if (layer == "u" && player.u.points.lt(320)) {
@@ -61,14 +62,15 @@ function getResetGain(layer, canMax=false, useType = null) {
 		if (player.u.points.gte(30) && layer == "u") resetgain = gain.floor().sub(player[layer].points).add(1).max(1)
 		return resetgain;
 	} else if (type=="normal"){
-		if (tmp[layer].baseAmount.lt(tmp[layer].requires)) return new Decimal(0)
+		if (tmp[layer].baseAmount.lt(tmp[layer].requires)) return decimalZero
 		let gain = tmp[layer].baseAmount.div(tmp[layer].requires).pow(tmp[layer].exponent).times(tmp[layer].gainMult).pow(tmp[layer].gainExp)
 		if (gain.gte(tmp[layer].softcap)) gain = gain.pow(tmp[layer].softcapPower).times(tmp[layer].softcap.pow(decimalOne.sub(tmp[layer].softcapPower)))
+		gain = gain.times(tmp[layer].directMult)
 		return gain.floor().max(0);
 	} else if (type=="custom"){
 		return layers[layer].getResetGain()
 	} else {
-		return new Decimal(0)
+		return decimalZero
 	}
 }
 
@@ -91,7 +93,7 @@ function getNextAt(layer, canMax=false, useType = null) {
 		if (!tmp[layer].canBuyMax) canMax = false
 		let base = new Decimal(1e8)
 		let exp = new Decimal(1.9)
-		let amt = player[layer].points.plus((canMax&&tmp[layer].baseAmount.gte(tmp[layer].nextAt))?tmp[layer].resetGain:0)
+		let amt = player[layer].points.plus((canMax&&tmp[layer].baseAmount.gte(tmp[layer].nextAt))?tmp[layer].resetGain:0).div(tmp[layer].directMult)
 		let g = amt
 		if (layer == "e") g = g.div(tmp.e.infDiv)
 		if (layer == "r" || layer == "u" || layer == "e") amt = scaleStaticCost(g, tmp[layer].row)
@@ -122,7 +124,7 @@ function getNextAt(layer, canMax=false, useType = null) {
 		if (tmp[layer].roundUpCost) cost = cost.ceil()
 		return cost;
 	} else if (type=="normal"){
-		let next = tmp[layer].resetGain.add(1)
+		let next = tmp[layer].resetGain.add(1).div(tmp[layer].directMult)
 		if (next.gte(tmp[layer].softcap)) next = next.div(tmp[layer].softcap.pow(decimalOne.sub(tmp[layer].softcapPower))).pow(decimalOne.div(tmp[layer].softcapPower))
 		next = next.root(tmp[layer].gainExp).div(tmp[layer].gainMult).root(tmp[layer].exponent).times(tmp[layer].requires).max(tmp[layer].requires)
 		if (tmp[layer].roundUpCost) next = next.ceil()
@@ -130,7 +132,7 @@ function getNextAt(layer, canMax=false, useType = null) {
 	} else if (type=="custom"){
 		return layers[layer].getNextAt(canMax)
 	} else {
-		return new Decimal(0)
+		return decimalZero
 	}}
 
 function softcap(value, cap, power = 0.5) {
@@ -229,7 +231,7 @@ function layerDataReset(layer, keep = []) {
 function resetBuyables(layer){
 	if (layers[layer].buyables) 
 		player[layer].buyables = getStartBuyables(layer)
-	player[layer].spentOnBuyables = new Decimal(0)
+	player[layer].spentOnBuyables = decimalZero
 }
 
 
@@ -280,7 +282,8 @@ function doReset(layer, force=false) {
 					if (!player[lrs[lr]].unlocked) player[lrs[lr]].unlockOrder++
 			}
 		}
-		tmp[layer].baseAmount = new Decimal(0) // quick fix
+	
+		tmp[layer].baseAmount = decimalZero // quick fix
 	}
 	if (tmp[layer].resetsNothing) return
 
@@ -289,7 +292,7 @@ function doReset(layer, force=false) {
 	}
 
 	prevOnReset = {...player} //Deep Copy
-	player.points = (row == 0 ? new Decimal(0) : getStartPoints())
+	player.points = (row == 0 ? decimalZero : getStartPoints())
 
 
 	for (let x = row; x >= 0; x--) rowReset(x, layer)
@@ -373,6 +376,7 @@ function completeChallenge(layer, x) {
 	let completions = canCompleteChallenge(layer, x)
 	if (!completions){
 		 player[layer].activeChallenge = null
+		 run(layers[layer].challenges[x].onExit, layers[layer].challenges[x])
 		return
 	}
 	if (player[layer].challenges[x] < tmp[layer].challenges[x].completionLimit) {
@@ -382,6 +386,7 @@ function completeChallenge(layer, x) {
 		if (layers[layer].challenges[x].onComplete) run(layers[layer].challenges[x].onComplete, layers[layer].challenges[x])
 	}
 	player[layer].activeChallenge = null
+	run(layers[layer].challenges[x].onExit, layers[layer].challenges[x])
 	updateChallengeTemp(layer)
 }