+++ js/technical/layerSupport.js
@@ -156,6 +156,15 @@ function setupLayer(layer){
         }  
     }
     
+    if (layers[layer].grid) {
+        layers[layer].grid.layer = layer
+        if (layers[layer].grid.getUnlocked === undefined)
+            layers[layer].grid.getUnlocked = true
+        if (layers[layer].grid.getCanClick === undefined)
+            layers[layer].grid.getCanClick = true
+
+    }
+
     if (layers[layer].startData) {
         data = layers[layer].startData()
         if (data.best !== undefined && data.showBest === undefined) layers[layer].showBest = true
@@ -165,8 +174,9 @@ function setupLayer(layer){
     if(!layers[layer].componentStyles) layers[layer].componentStyles = {}
     if(layers[layer].symbol === undefined) layers[layer].symbol = layer.charAt(0).toUpperCase() + layer.slice(1)
     if(layers[layer].unlockOrder === undefined) layers[layer].unlockOrder = []
-    if(layers[layer].gainMult === undefined) layers[layer].gainMult = new Decimal(1)
-    if(layers[layer].gainExp === undefined) layers[layer].gainExp = new Decimal(1)
+    if(layers[layer].gainMult === undefined) layers[layer].gainMult = decimalOne
+    if(layers[layer].gainExp === undefined) layers[layer].gainExp = decimalOne
+    if(layers[layer].directMult === undefined) layers[layer].directMult = decimalOne
     if(layers[layer].type === undefined) layers[layer].type = "none"
     if(layers[layer].base === undefined || layers[layer].base <= 1) layers[layer].base = 2
     if(layers[layer].softcap === undefined) layers[layer].softcap = new Decimal("e1e7")