+++ js/technical/temp.js
@@ -6,10 +6,12 @@ var NaNalert = false;
 // Tmp will not call these
 var activeFunctions = [
 	"startData", "onPrestige", "doReset", "update", "automate",
-	"buy", "buyMax", "respec", "onComplete", "onPurchase", "onPress", "onClick", "onHold", "masterButtonPress",
+	"buy", "buyMax", "respec", "onPress", "onClick", "onHold", "masterButtonPress",
 	"sellOne", "sellAll", "pay", "actualCostFunction", "actualEffectFunction",
 	"effectDescription", "display", "fullDisplay", "effectDisplay", "rewardDisplay",
 	"tabFormat", "content",
+	"onComplete", "onPurchase", "onEnter", "onExit",
+	"getUnlocked", "getStyle", "getCanClick", "getTitle", "getDisplay"
 ]
 
 var noCall = doNotCallTheseFunctionsEveryTick
@@ -41,19 +43,11 @@ function setupTemp() {
 	}
 
 	tmp.other = {
-		screenWidth: window.innerWidth,
-		splitScreen: window.innerWidth >=1024,
-		lastPoints: player.points || new Decimal(0),
-		oomps: new Decimal(0),
-
-		held: {
-			time: null,
-			id: null,
-			layer: null,
-			type: null,
-		}
+		lastPoints: player.points || decimalZero,
+		oomps: decimalZero,
     }
 
+	updateWidth()
 
 	temp = tmp
 }
@@ -81,7 +75,7 @@ function setupTempData(layerData, tmpData, funcsData) {
 		}
 		else if (isFunction(layerData[item]) && !activeFunctions.includes(item)){
 			funcsData[item] = layerData[item]
-			tmpData[item] = new Decimal(1) // The safest thing to put probably?
+			tmpData[item] = decimalOne // The safest thing to put probably?
 		} else {
 			tmpData[item] = layerData[item]
 		}