+++ js/utils.js
@@ -91,7 +91,15 @@ function powExp3(n, exp){
 	return Decimal.pow(10,Decimal.pow(10,Decimal.pow(10,n.log10().log10().log10().pow(exp))))
 }
 
-function upgradeEffect(layer, id){
+function getGridData(layer, id) {
+	return (player[layer].grid[id])
+}
+
+function setGridData(layer, id, data) {
+	player[layer].grid[id] = data
+}
+
+function upgradeEffect(layer, id) {
 	return (tmp[layer].upgrades[id].effect)
 }
 
@@ -114,6 +122,11 @@ function clickableEffect(layer, id) {
 function achievementEffect(layer, id) {
 	return (tmp[layer].achievements[id].effect)
 }
+
+function gridEffect(layer, id) {
+	return (gridRun(layer, 'getEffect', player[layer].grid[id], id))
+}
+
 
 function canAffordPurchase(layer, thing, cost) {
 
@@ -204,6 +217,14 @@ function clickClickable(layer, id) {
 
 	run(layers[layer].clickables[id].onClick, layers[layer].clickables[id])
 	updateClickableTemp(layer)
+}
+
+function clickGrid(layer, id) {
+	if (!player[layer].unlocked) return
+	if (!run(layers[layer].grid.getUnlocked, layers[layer].grid, id)) return
+	if (!gridRun(layer, 'getCanClick', player[layer].grid[id], id)) return
+
+	gridRun(layer, 'onClick', player[layer].grid[id], id)
 }
 
 // Function to determine if the player is in a challenge
@@ -468,4 +489,13 @@ function run(func, target, args = null) {
 	}
 	else
 		return func;
+}
+
+function gridRun(layer, func, data, id) {
+	if (isFunction(layers[layer].grid[func])) {
+		let bound = layers[layer].grid[func].bind(layers[layer].grid)
+		return bound(data, id)
+	}
+	else
+		return layers[layer].grid[func];
 }