+++ changelog.md
@@ -1,5 +1,23 @@
 # The Modding Tree changelog:
 
+# v2.5.5.1 - 5/12/21
+- Fixed clickables.
+
+# v2.5.5 - 5/12/21
+- Added grids! They are a grid of buttons which behave the same, but have their own data. Good for inventory grids, map tiles, and more!
+- Added "marked" feature to add a mark to a node. Can be an image instead of a star. (Originally by Jacorb)
+- Added "layer-proxy" component that lets you use components from another layer.
+- Added the ability to display non-whole numbers in main-display.
+
+# v2.5.4 - 5/10/21
+- Added a setting to always use single-tab mode.
+- Added directMult, which multiplies prestige gain after exponents and softcaps. It actually multiplies gain for static layers.
+- Added onEnter and onExit for challenges.
+- Improved displaying numbers between 0.0001 and 0.1.
+- Added documentation on how gainMult/Exp work for static layers.
+- Fixed a visual issue on mobile, thanks to thepaperpilot.
+- Improved documentation in general.
+
 # v2.5.3 - 5/8/21
 - Improved performance of tab formats and bars.
 - Respec confirmation settings are now kept on resets.